Add fix-git-locale that deletes `es` locale in git
Git has a problem with it's locale when it's installed with Homebrew.
There are several ways to go and fix it and one of them is just to delete
the Spanish locale in the Git installation folder. It's to be done every
time you install Git. This command address this problem.